requirements-api.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
SQLAlchemy>=2.0.43
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
pydantic>=2.11.7
pydantic-settings>=2.10.1
email-validator>=2.3.0
sendgrid>=6.12.5
